A  blog plugin.
The plugin heavily depends on the existing framework for its job.
So the integration is not yet complete.

Everything is command based and does not work on its own. Apart from a
dummy cron. The cron rotates every 3 hours, and asks a question.

The regex is pretty generic and swallows everyghin thrown at it. The
mathced string is stored as it it.