cron with thread sleep in waiting time.
So the threadpool can also schedule tasks. The granurality of the
scheduler is not very high but that is indenpendent of the pool.

The get either an interval to repeat the job, or the interval can be
decided by the job itself.

Threadpool delays scheduling of the job by staring a waiter thread,
which sleeps for the waiting time (decided by the checker passed by
job), and then schedules the actual job of execution.