new plugin sleeper.
A debug plugin, I see time lags in rpi. It appears that the pi corrects
its time after boot, this can be due to unplanned shutdown occuring at
power loss. Since rpi corrects itself and the cron goes outofsync.

In that case a fix will be required in the cron module.