reload the crons; based on command.
The rpi ntpd updates the time causing the crons go out of sync,
this reload mechanism, tracks and informs when the time goes out of
sync. Requiring the user to restart the bot. This stops all threads.

Scheduled threads are stopped by a threading.event(), then all these
threads come back to waiting on the queue. then a special function it
added to task which raises a StopThread exception making the threads
return.

Thus all the threads are stopped.