reduced logging mess; change in cron job registering paths.
The jobs registering happens only when the plugin is created and a call
is made ot register jobs. In that call crontable was made explicitly
empty so subsequent calls did not register jobs. However the call to
register jobs was always made. This is not requered hence removed one
call.