scheduled functions can specify maximum number of parallel jobs.
Plugin class now has a maxcounts dicts with key as id of the function
being scheduled and schedule timer function is co-routined to check if
any more parallel executions can be done.

Also now the threadpool if receives a negative value will wait the
for the  value time but since the value was negative it will now add it
as task rather ask for next wait time.