[youtube] Progress hook integration for proper return value to slack interface (#3)
* Temporary global variable solution to return value of downloader function

A temporary global variable has been introduced in this case. However,
this method does not work well for audio conversions as the return value
is still the name of the video. So need to find a solution to this.
Also, the variable need to be made a part of the thread class so that it
stays implicit.

* A temp commit for error outputted by the bot to slack

* Done with setting up downloading_list list in every possibility. plgn.new and plgn.old no more needed.