[recron] added functionality to keep cron in conifg
crons for msg handlers can now be set in config.
as
`
someplgn:
       msghandler:cronstring
`
There still a chance that the cron does not work as desired. But that
should be a rare case or the cronstring is too complicated.

Significant change in architecture of the `Plugin` Class. Also there are
a couple of hack.
So overall the stability is questionable.